Move NIP-10 thread references into protocol
